Added support for odd-dimensioned images, other refactoring
Previously, images had to be multiples of 8, now they can be any dimensions.

Additionally, this commit refactors the use of "exit" in AtriumCore
to instead throw various exceptions that can be caught in the main.
And it includes lots of name refactoring.

Fixed the ordering of the setting bits in the zig-zag pattern
to follow that used by the JPEG standard pattern.

Encode, decode works for a small message with some of the test images.
It looks like the program has issues with white backgrounds.