Added JPEG ZIG ZAG ordering and flexible encoding start position
The full JPEG ZIG ZAG ordering is defined and added.
The cells used to encode the bits of a byte are based on that ordering,
and the starting position is exposed as an argument.

This also updates the AtriumSteganographySpec
which was broken in the last commit's changes to BIT CELLS.