Fix YCC range values and add DCT quantized correction factor
This commit fixes a significant bug in the handling of Y Luminance values.
Previously, when converting back from a DCT, the luminance could exceed
its valid range of values. This would not fail when converting into RGB,
but it would fail when decoding from that RGB into YCC.

The fix is to clamp the Y luminance value to its valid range.

In addition, this commit adds a small quantization correction factor
to move DCT coefficients that are on the border of a quantized value.
This prevents JPEG artifacts from changing the coefficient's quantized parity.

Lastly, this commit refactors a lot of code and deletes unused code.